refactor generated code
